Add files via upload
To turn on an external LED, we selected pin PC0 as an output and configured it as "GPIO_Output." The blue button PB1 remains configured as an input. Next, we implemented the setup on a prototype board using a red LED and a 1kΩ resistor.

In this setup, pin PC0 is connected to the anode of the LED (the longer side). The LED is in series with the resistor, which is connected to the ground (GND) of the board.