Add files via upload
The external LEDs can be controlled using interrupts, which involve interrupting the current operation of the microcontroller to execute our code. To achieve this, we configure the blue button PB1 by defining PA0 as "GPIO_EXTI0." In this exercise, we used LED3 as the output. Additionally, in the "Configuration" menu, it is necessary to check "EXTI line0 interrupt" in the "NVIC Configuration" window, and set the GPIO level to "LOW" in the "Pin Configuration" window. This is because in the interrupt, we work with a low level.