Add files via upload
To control the internal LEDs, it is necessary to configure them as inputs and the push-button as an output. To achieve this, right-click in the box corresponding to the appropriate pin.

For instance, to define the blue button PB1 as an input, you need to set the PA0 box as "GPIO_Input" in the selection menu.